Add optional CLI operations
Add a second path for using commandline arguements when CLI mode is
active. The climode is activated when the module 'tikinter' is not
found. Within the CLI mode the script expects at least one argument as
filename to the original MS TMT file.

The output is a json file with the same basename. This is already
handled by the original code.